+++ casey-git-clone/src/casey_git_clone/__init__.py
@@ -1,5 +1,7 @@
 """Casey Git Clone core package."""
 
+from .exchange import CaseyOverlayRefs, casey_to_sb_overlay_record
+from .ledger_sqlite import BuildLedgerRecord, OperationLedgerRecord
 from .models import (
     Blob,
     BuildView,
@@ -25,4 +27,8 @@ __all__ = [
     "collapse_conflict",
     "publish_edits",
     "sync_workspace",
+    "OperationLedgerRecord",
+    "BuildLedgerRecord",
+    "CaseyOverlayRefs",
+    "casey_to_sb_overlay_record",
 ]